Extract some code from `Expression.kt`
